# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger KW-API",
        "description": "This is KW-API messaging middleware for dispatch data to services / applications",
        "contact": {
            "email": "izhur2001@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api_users": {
            "get": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "list of api users",
                "description": "",
                "operationId": "list api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page data",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "Create api user",
                "description": "",
                "operationId": "create api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "apiKey",
                        "in": "formData",
                        "description": "apiKey of api user",
                        "required": true,
                        "type": "string",
                        "default": "1234567890abcdef"
                    },
                    {
                        "name": "company",
                        "in": "formData",
                        "description": "company user of api key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "formData",
                        "description": "application name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api_users/{id}": {
            "get": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "show detail of api user",
                "description": "",
                "operationId": "get api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of api user",
                        "required": true,
                        "type": "integer",
                        "default": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "Update api user data",
                "description": "",
                "operationId": "update api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of api user",
                        "required": true,
                        "type": "integer",
                        "default": "1"
                    },
                    {
                        "name": "apiKey",
                        "in": "formData",
                        "description": "apiKey of api user",
                        "required": true,
                        "type": "string",
                        "default": "1234567890abcdef"
                    },
                    {
                        "name": "company",
                        "in": "formData",
                        "description": "company user of api key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "formData",
                        "description": "application name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "Delete api user",
                "description": "",
                "operationId": "delete api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of api user",
                        "required": true,
                        "type": "integer",
                        "default": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api_users/{id}/toggle": {
            "get": {
                "tags": [
                    "apiUsers"
                ],
                "summary": "Toggle active status of api user",
                "description": "",
                "operationId": "toggle api_user",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of api user",
                        "required": true,
                        "type": "integer",
                        "default": "1"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiKey": {
            "type": "apiKey",
            "name": "apiKey",
            "in": "header"
        }
    },
...
}